improve sockets PR FIX
