+++ client/src/components/ActiveChat/Input.js
@@ -38,7 +38,7 @@ class Input extends Component {
       text: event.target.text.value,
       recipientId: this.props.otherUser.id,
       conversationId: this.props.conversationId,
-      sender: this.props.conversationId ? this.props.user : null,
+      sender: this.props.conversationId ? null : this.props.user,
       read: false,
     };
     await this.props.postMessage(reqBody);
